# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Supply Chain Finance
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /demandorder:
    x-swagger-router-controller: DemandOrder
    get:
      tags:
        - Demand Order
      description: Returns demand order details
      # used as the method name of the controller
      operationId: getDemandOrder
      parameters:
        - name: doNumber
          description: Demand order Number
          type: integer
          in: query
          required: false
        - name: sortBy
          in: query
          description: sort by query
          required: false
          type: string
        - name: searchBy
          in: query
          description: search by query
          required: false
          type: string
        - name: page
          in: query
          description: pagination query
          required: false
          type: integer
      responses:
        200:
          description: Success
          schema:
            type: array
            items:
              # a pointer to a definition
              $ref: "#/definitions/DemandOrder"
        201:
          description: "Created"
        304:
          description: "Not Modified"
          schema:
            $ref: "#/definitions/ErrorResponse"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/ErrorResponse"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/ErrorResponse"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/ErrorResponse"
        417:
          description: "Expectation Failed"
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
        503:
          description: "Service Unavailable"
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      tags:
        - Demand Order
      description: Delete demand order details
      # used as the method name of the controller
      operationId: deleteDemandOrder
      parameters:
        - name: doNumber
          description: Demand order Number
          type: integer
          in: query
          required: true
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/ApiResponse'
        201:
          description: "Created"
        304:
          description: "Not Modified"
          schema:
            $ref: "#/definitions/ErrorResponse"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/ErrorResponse"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/ErrorResponse"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/ErrorResponse"
        417:
          description: "Expectation Failed"
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
        503:
          description: "Service Unavailable"
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      tags:
        - Demand Order
      description: add new demand order detail
      operationId: createDemandOrder
      parameters:
        - in: body
          name: body
          description: Demand Order details
          required: true
          schema:
              # a pointer to a definition
              $ref: "#/definitions/DemandOrder"
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/ApiResponse'
        201:
          description: "Created"
        304:
          description: "Not Modified"
          schema:
            $ref: "#/definitions/ErrorResponse"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/ErrorResponse"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/ErrorResponse"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/ErrorResponse"
        417:
          description: "Expectation Failed"
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
        503:
          description: "Service Unavailable"
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      tags:
        - Demand Order
      description: update demand order details
      operationId: updateDemandOrder
      parameters:
        - in: body
          name: body
          description: update demand order details
          required: true
          schema:
            $ref: "#/definitions/UpdateDemandOrder"
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/ApiResponse'
        201:
          description: "Created"
        304:
          description: "Not Modified"
          schema:
            $ref: "#/definitions/ErrorResponse"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/ErrorResponse"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/ErrorResponse"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/ErrorResponse"
        417:
          description: "Expectation Failed"
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
        503:
          description: "Service Unavailable"
          schema:
            $ref: "#/definitions/ErrorResponse"
    patch:
      tags:
        - Demand Order
      description: update single entity of Demand Order
      operationId: updateSingleDemandOrder
      parameters:
        - in: body
          name: body
          description: update demand order details
          required: true
          schema:
            $ref: "#/definitions/UpdateDemandOrder"
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/ApiResponse'
        201:
          description: "Created"
        304:
          description: "Not Modified"
          schema:
            $ref: "#/definitions/ErrorResponse"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/ErrorResponse"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/ErrorResponse"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/ErrorResponse"
        417:
          description: "Expectation Failed"
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
        503:
          description: "Service Unavailable"
          schema:
            $ref: "#/definitions/ErrorResponse"
  /org:
    x-swagger-router-controller: Organization
    get:
      tags:
        - Organizations
      description: Returns Organization details
      # used as the method name of the controller
      operationId: getOrganization
      parameters:
        - name: orgId
          in: query
          description: Organization ID
          required: false
          type: integer
        - name: sortBy
          in: query
          description: sort by query
          required: false
          type: string
        - name: searchBy
          in: query
          description: search by query
          required: false
          type: string
        - name: page
          in: query
          description: pagination query
          required: false
          type: integer
      responses:
        200:
          description: Success
          schema:
            type: array
            items:
              # a pointer to a definition
              $ref: "#/definitions/organization"
        201:
          description: "Created"
        304:
          description: "Not Modified"
          schema:
            $ref: "#/definitions/ErrorResponse"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/ErrorResponse"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/ErrorResponse"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/ErrorResponse"
        417:
          description: "Expectation Failed"
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
        503:
          description: "Service Unavailable"
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      tags:
        - Organizations
      description: Delete organization details
      # used as the method name of the controller
      operationId: deleteOrganization
      parameters:
        - name: orgId
          in: query
          description: Organization ID
          required: true
          type: integer
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/ApiResponse'
        201:
          description: "Created"
        304:
          description: "Not Modified"
          schema:
            $ref: "#/definitions/ErrorResponse"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/ErrorResponse"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/ErrorResponse"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/ErrorResponse"
        417:
          description: "Expectation Failed"
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
        503:
          description: "Service Unavailable"
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      tags:
        - Organizations
      description: add new Organization details
      operationId: createOrganization
      parameters:
        - in: body
          name: body
          description: Organization details
          required: true
          schema:
              # a pointer to a definition
              $ref: "#/definitions/organization"
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/ApiResponse'
        201:
          description: "Created"
        304:
          description: "Not Modified"
          schema:
            $ref: "#/definitions/ErrorResponse"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/ErrorResponse"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/ErrorResponse"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/ErrorResponse"
        417:
          description: "Expectation Failed"
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
        503:
          description: "Service Unavailable"
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      tags:
        - Organizations
      description: update Organization details
      operationId: updateOrganization
      parameters:
        - in: body
          name: body
          description: update demand order details
          required: true
          schema:
            $ref: "#/definitions/UpdateOrganization"
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/ApiResponse'
        201:
          description: "Created"
        304:
          description: "Not Modified"
          schema:
            $ref: "#/definitions/ErrorResponse"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/ErrorResponse"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/ErrorResponse"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/ErrorResponse"
        417:
          description: "Expectation Failed"
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
        503:
          description: "Service Unavailable"
          schema:
            $ref: "#/definitions/ErrorResponse"
    patch:
      tags:
        - Organizations
      description: update Organization details
      operationId: updateSingleOrganization
      parameters:
        - in: body
          name: body
          description: update demand order details
          required: true
          schema:
            $ref: "#/definitions/UpdateOrganization"
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/ApiResponse'
        201:
          description: "Created"
        304:
          description: "Not Modified"
          schema:
            $ref: "#/definitions/ErrorResponse"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/ErrorResponse"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/ErrorResponse"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/ErrorResponse"
        417:
          description: "Expectation Failed"
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
        503:
          description: "Service Unavailable"
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  UpdateOrganization:
    type: object
    properties:
      condition:
        $ref: "#/definitions/orgCondition"
      update:
        $ref: "#/definitions/organization"
    xml:
      name: "UpdateOrganization"
  orgCondition:
    type: "object"
    required:
    - "orgId"
    properties:
      orgId:
        type: "integer"
        format: "int64"
    xml:
      name: "orgCondition"
  UpdateDemandOrder:
    type: object
    properties:
      condition:
        $ref: "#/definitions/doCondition"
      update:
        $ref: "#/definitions/DemandOrder"
    xml:
      name: "UpdateDemandOrder"
  doCondition:
    type: "object"
    required:
    - "doNumber"
    properties:
      doNumber:
        type: "integer"
        format: "int64"
    xml:
      name: "doCondition"
...
  DemandOrder:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      manufacturerName:
        type: "string"
      doNumber:
        type: "integer"
        format: "int64"
      createDate:
        type: "string"
        format: "date-time"
      issueDate:
        type: "string"
        format: "date-time"
      distributorName:
        type: "string"
      addressOfDistributor:
        type: "string"
      deliveryAddress:
        type: "string"
      billingAddress:
        type: "string"
      product:
        type: "string"
      description:
        type: "string"
      unit:
        type: "string"
      totalPrice:
        type: "integer"
        format: "int32"
      grandTotal:
        type: "integer"
        format: "int32"
      advancePayment:
        type: "integer"
        format: "int32"
      quantity:
        type: "integer"
        format: "int32"
    example:
      product: "product"
      quantity: 2
      totalPrice: 1
      grandTotal: 5
      description: "description"
      manufacturerName: "nameOfmanufacturer"
      doNumber: 6
      unit: "unit"
      deliveryAddress: "deliveryAddress"
      advancePayment: 5
      addressOfDistributor: "addressOfDistributor"
      id: 0
      billingAddress: "billingAddress"
      issueDate: "2000-01-23T04:56:07.000+00:00"
      createDate: "2000-01-23T04:56:07.000+00:00"
      distributorName: "nameOfDistributor"
    xml:
      name: "DemandOrder"
  organization:
    type: object
    properties:
      orgName:
        type: string
        description: organization name
      orgId:
        type: integer
        format: int64
        description: organization id
    xml:
      name: organization
...
  ApiResponse:
    type: object
    properties:
      code:
        type: integer
        format: int32
      type:
        type: string
      message:
        type: string
    example:
      code: 0
      type: type
      message: message
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string